+++ Model/Model/TourLog.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Text.Json.Serialization;
 
 namespace Model
 {
@@ -9,7 +8,7 @@ namespace Model
         public DateTime Date { set; get; } 
         public Type Type { set; get; }
         public TimeSpan Duration { set; get; }
-        public int Distance { set; get; }
+        public double Distance { set; get; }
         public int Rating { set; get; }
         public string Report { set; get; }
         public double AvgSpeed { set; get; }
@@ -17,7 +16,7 @@ namespace Model
         public double HeightDifference { set; get; }
         public int Stops { set; get; }
 
-        public TourLog(DateTime date, Type type, TimeSpan duration, int distance, int rating, string report, double avgSpeed, double maxSpeed, double heightDifference, int stops)
+        public TourLog(DateTime date, Type type, TimeSpan duration, double distance, int rating, string report, double avgSpeed, double maxSpeed, double heightDifference, int stops)
         {
             Date = date;
             Type = type;
@@ -31,7 +30,7 @@ namespace Model
             Stops = stops;
         }
         
-        public TourLog(int id, DateTime date, Type type, TimeSpan duration, int distance, int rating, string report, double avgSpeed, double maxSpeed, double heightDifference, int stops)
+        public TourLog(int id, DateTime date, Type type, TimeSpan duration, double distance, int rating, string report, double avgSpeed, double maxSpeed, double heightDifference, int stops)
         {
             Id = id;
             Date = date;
