+++ Server/Server/DAL/IMapApiResponse.cs
@@ -2,11 +2,11 @@
 {
     public class MapApiResponse
     {
-        public int Distance { get; }
+        public double Distance { get; }
         public string BoundingBox { get; }
         public string SessionId { get; }
         
-        public MapApiResponse(int distance, string boundingBox, string sessionId)
+        public MapApiResponse(double distance, string boundingBox, string sessionId)
         {
             Distance = distance;
             BoundingBox = boundingBox;
